api/pyproject.toml
[project]
name = "sample-agent"
version = "0.1.0"
description = "A LangGraph agent"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.2.0",
    "langgraph==1.0.5",
    "langsmith>=0.4.49",
    "openai>=1.68.2,<2.0.0",
    "fastapi>=0.115.5,<1.0.0",
    "uvicorn>=0.29.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "langchain-openai>=1.1.0",
    "copilotkit>=0.1.74",
    "langgraph-cli[inmem]>=0.1.55",
]
